Python3 anuga py3

+++ anuga/file/mux.py
@@ -1,7 +1,9 @@
 """
     Read a mux2 file.
 """
+from __future__ import absolute_import
 
+from builtins import range
 from anuga.utilities.numerical_tools import ensure_numeric
 import numpy as num     
         
@@ -35,7 +37,7 @@ def read_mux2_py(filenames,
                         extracted
     """
 
-    from urs_ext import read_mux2
+    from .urs_ext import read_mux2
 
     numSrc = len(filenames)
 
@@ -51,7 +53,7 @@ def read_mux2_py(filenames,
         weights = num.ones(numSrc)
 
     if permutation is None:
-        permutation = ensure_numeric([], num.float)
+        permutation = ensure_numeric([], num.int)
 
     # Call underlying C implementation urs2sts_ext.c
     data = read_mux2(numSrc, filenames, weights, file_params,
